Moved mapcomponent under plotsearchAndCompetitions/components. Fixed css so that both mapSearchComponent and mapComponent are visible correctly
