--- src/map/_map-view.scss
@@ -1,6 +1,0 @@
-/* placeholder for custom map css */
-.leaflet-container {
-  position: absolute;
-  height: 100%;
-  width: 100%;
-}
