+++ src/plotSearchAndCompetitionsPage/components/_mapComponent.scss
@@ -1,0 +1,4 @@
+.mapComponent {
+  height: 100vh;
+  z-index: 10;
+}
